refactor(build): flatten /config and /build
