--- config/development.js
@@ -1,5 +1,0 @@
-module.exports = (config) => {
-  return {
-    compiler_devtool: 'eval-cheap-module-source-map',
-  }
-}
