--- config/test.js
@@ -1,4 +1,0 @@
-// Just use the production configuration
-const production = require('./production')
-
-module.exports = production
