+++ gulp/tasks/docs.js
@@ -31,7 +31,7 @@ task('generate-docs-json', () => {
 })
 
 task('webpack-docs', (cb) => {
-  const webpackConfig = require('../../build/webpack.config')
+  const webpackConfig = require('../../webpack.config')
   const compiler = webpack(webpackConfig)
 
   compiler.run((err, stats) => {
