+++ gulp/tasks/serve.js
@@ -12,7 +12,7 @@ const g = loadPlugins()
 const { log, colors } = g.util
 
 const serve = (cb) => {
-  const webpackConfig = require('../../build/webpack.config')
+  const webpackConfig = require('../../webpack.config')
   const app = express()
   const compiler = webpack(webpackConfig)
