+++ package.json
@@ -26,7 +26,7 @@
     "start": "npm run docs",
     "start:local-modules": "npm run docs -- --local-modules",
     "pretest": "gulp dll",
-    "test": "cross-env NODE_ENV=test karma start build/karma.conf.babel.js",
+    "test": "cross-env NODE_ENV=test karma start",
     "test:watch": "npm run test --silent -- --watch"
   },
   "repository": {
